1. Archivo de código y documento/Archivo/requirements.txt
django==4.1.5
django-environ==0.9.0 
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2==2.9.5
